Renaming duns.py to sam_recipient.py
